backend/database.py
import os
from supabase import create_client, Client
from dotenv import load_dotenv
import json

load_dotenv()

supabase_url = os.environ.get("SUPABASE_URL")
supabase_key = os.environ.get("SUPABASE_KEY")

supabase: Client = create_client(supabase_url, supabase_key)