backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

# HTTP requests (para scraping y llamadas a n8n / APIs)
requests==2.32.3
httpx==0.27.0

# Base de datos (PostgreSQL / Supabase)
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# Validación de datos
pydantic==2.7.1

# Manejo de variables de entorno
python-dotenv==1.0.1

# Parsing y scraping
beautifulsoup4==4.12.3
lxml==5.2.2

# Utilidades
python-dateutil==2.9.0

# Supabase
supabase==2.28.3